refactor: use fixed pre-range lead options (1-3C)
Replace free-form pre-range lead input with a simple 1/2/3C dropdown and clamp values in settings normalization. This keeps thermal trigger behavior predictable in simple mode and avoids overly fine tuning.

+++ NINA.Plugin.SeeDark/SeeDarkSettings.cs
@@ -18,7 +18,7 @@ namespace NINA.Plugin.SeeDark {
         public string DiscordWebhookUrl { get; set; } = "";
         public int TempBucketSize { get; set; } = 2;
         public int StackTolerance { get; set; } = 2;
-        public double PreBucketLeadC { get; set; } = 1.0;
+        public int PreBucketLeadC { get; set; } = 1;
 
         private static string SettingsPath => Path.Combine(DataFolder, "settings.json");
